rdc_video_bot/requirements.txt
google-auth-oauthlib
google-api-python-client
pandas
python-dotenv
gspread
gspread-dataframe